+++ application/controllers/About.php
@@ -2,25 +2,26 @@
 
 defined('BASEPATH') OR exit('No direct script access allowed');
 
+
+
+/*
+ * This class represents the about page of our airport info site.
+ * Currently empty.
+ * 
+ * @author Jonathan
+ */
 class About extends Application
 {
 
-	/**
-	 * Index Page for this controller.
-	 *
-	 * Maps to the following URL
-	 * 		http://example.com/
-	 * 	- or -
-	 * 		http://example.com/welcome/index
-	 *
-	 * So any other public methods not prefixed with an underscore will
-	 * map to /welcome/<method_name>
-	 * @see https://codeigniter.com/user_guide/general/urls.html
-	 */
+        /**
+        * About page for the application explaining the purpose of this website.
+        *
+        * Maps to the following URL
+        * 		http://comp4711.local/about
+        */
 	public function index()
 	{
-		$this->data['pagebody'] = 'about';
-                
+		$this->data['pagebody'] = 'about';        
 		$this->render();
 	}