+++ application/controllers/Homepage.php
@@ -2,69 +2,56 @@
 
 defined('BASEPATH') OR exit('No direct script access allowed');
 
+/*
+ * This class represents the homepage of our airport info site.
+ * @author Sergey
+ */
 class Homepage extends Application
 {
 
-	/**
-	 * Index Page for this controller.
-	 *
-	 * Maps to the following URL
-	 * 		http://example.com/
-	 * 	- or -
-	 * 		http://example.com/welcome/index
-	 *
-	 * So any other public methods not prefixed with an underscore will
-	 * map to /welcome/<method_name>
-	 * @see https://codeigniter.com/user_guide/general/urls.html
-	 */
-        public function array_icount_values($array) {
-            //create new output array of style date : # occurences
-            //cycle through given dates, initializing to date : 1 except if duplicate
-            //if duplicate, add +1 to occurence
-            //return array
-            
-            $outarray = array();
-            //array(	'0'	 => array('id' => '666', 'make'	 => 'Boeing')
-
-            $count = -1;
-            foreach ($array as $key=>$value) {
-                if (!isset($outarray[$value['date']])) {
-                    
-                    $outarray[++$count] = array('date' => $value['date'], 
-                        'count' => 1);
-                } else {
-                    foreach ($outarray as $key2 => $value2 ) {
-                        if ($value2['date'] == $value['date']) {
-                            $value2['count'] += 1;
-                        }
-                    }
-                }
+        /*
+         * Function that takes in an array of values and dates, then
+         * returns an array with dates and the number of occurences
+         * of each date.
+         */
+        public function countForDates($array)
+        {
+          $countForDate = array();
+          foreach ($array as $key=>$value) {
+            if (isset($countForDate[$value['date']])) {
+              $countForDate[$value['date']] += 1;
+            } else {
+              $countForDate[$value['date']] = 1;
             }
-            
-            return $outarray;
+          }
+
+          $result = [];
+          foreach ($countForDate as $key=>$value) {
+            // Build date element per codeigniter format.
+            $date = array('date'=>$key, 'count'=>$value);
+            // Append date.
+            $result[] = $date;
+          }
+
+          return $result;
         }
-        
-        
+
+
 	public function index()
 	{
-		$this->data['pagebody'] = 'homepage';
-                
-                // build the list of authors, to pass on to our view
-                $flightsrc = $this->flightsmdl->all();
-                
-                // pass on the data to present, as the "authors" view parameter
-                $this->data['flights'] = $flightsrc;
-                
-                
-                $this->data['datearr'] = $this->array_icount_values($flightsrc);
-                
-                //echo "DUMPING";
-                //var_dump($this->array_icount_values($flightsrc));
-                
-                
-		$this->render();
+            $this->data['pagebody'] = 'homepage';
+            // pass on the data to present, including flights, dates of flights,
+            // the size of hte fleet, the base airport, and destination airports
+            $flightsrc = $this->flightsmdl->all();
+            $this->data['flights'] = $flightsrc;
+            $this->data['datearr'] = $this->countForDates($flightsrc);
+            $this->data['sizeFleet'] = sizeof($this->fleetmdl->all());
+            $this->data['baseAirport'] = $this->flightsmdl->getBaseApt();
+            $this->data['airports'] = $this->flightsmdl->getDestApt();
+            
+            //display the page
+            $this->render();
+            
+            
 	}
-        
-
-
 }