+++ application/views/flights.php
@@ -1,8 +1,11 @@
 <h1>Flights</h1>
-<br /><br />
+<br/><br/>
 
 <table class="table">
     <tr>
+        <th>
+            Aircraft Code
+        </th>
         <th>
             From
         </th>
@@ -15,6 +18,11 @@
     </tr>  
     {flights}
     <tr>
+        <td>
+            <a style="text-decoration: none; color:black;" ref="#" title="The aircraft code is a unique identifier for the aircraft involved in a flight.">
+            {accode}
+            </a>
+        </td>
         <td>
             {from}
         </td>