+++ changelog.md
@@ -10,6 +10,17 @@ Project contributors:
 - Kuanysh
 - Jonathan
 
+## [Unreleased] - Oct 8, 2.36 pm
+### Updated
+- Renamed flights model to resolve bug.
+
+## [Unreleased] - Oct 8, 1.12 pm
+### Updated
+- Improved homepage with information about origin and destination of flights and number of planes.
+
+## [Unreleased] - Oct 5, 8.12 pm
+### Updated
+- Added info/fleet and info/flights controllers that display JSON data returned from models
 
 ## [Unreleased] - Oct 5, 7.56 pm
 ### Updated